% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matching_subjects}
\alias{matching_subjects}
\title{Propensity Score matching}
\usage{
matching_subjects(x)
}
\arguments{
\item{x}{A numeric vector with probabilities}
}
\value{
A list.
}
\description{
Propensity Score matching
}
\examples{
# Testing the function
x <- runif(10)
matching_subjects(x)
}
