% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Here is the function}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\description{
Here is the function
}
\examples{

#Here is an example
set.seed(312)
x <- rnorm(100)
y <- rnorm(100)
myplot(x, y)
}
