% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_match1}
\alias{ps_match1}
\title{psmatch from Rcpp}
\usage{
ps_match1(x)
}
\value{
A list with two vectors (\code{match_id}, \code{match_x}).
}
\description{
psmatch from Rcpp
}
\examples{
# Testing the function
set.seed(1231)
x <- cbind(runif(5))
}
