% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_matching}
\alias{ps_matching}
\title{Propesnity Score}
\usage{
ps_matching(x)
}
\value{
A list with two vectors (\code{match_id}, \code{match_x}).
}
\description{
Propesnity Score
}
\examples{
set.seed(1231)
x <- cbind(runif(5))
ps_matching(x)
}
