% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{Hello world from Rcpp}
\usage{
rcpp_hello_world()
}
\value{
A list with two vectors (\code{x}, \code{y}).
}
\description{
Hello world from Rcpp
}
\examples{
rcpp_hello_world()
}
