% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matching_subjects}
\alias{matching_subjects}
\title{Match subjects Rcpp}
\usage{
matching_subjects(x)
}
\arguments{
\item{x}{NumericVector. This is the input}
}
\value{
A list with two vectors (\code{match_id}, \code{match_x}).
}
\description{
Match subjects Rcpp
}
\examples{
# Testing the function
library(Rcpp)
x = c(1,2,3,4,5,6)
matching_subjects(x)
}
