% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_match}
\alias{ps_match}
\title{Ps matching from Rcpp}
\usage{
ps_match(x)
}
\arguments{
\item{x}{\itemize{
\item A vector to be used for the matching
}}
}
\value{
A list of matched results.
}
\description{
Ps matching from Rcpp
}
\examples{
set.seed(1231)
x <- cbind(runif(5))
ps_match(x)
}
