% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_match}
\alias{ps_match}
\title{Propensity score matching}
\usage{
ps_match(x)
}
\arguments{
\item{x}{Numeric vector with probablities.}
}
\value{
A list
}
\description{
Propensity score matching
}
\examples{
set.seed(7123)
x <- runif(10)
ps_match(x)
}
