% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_match3}
\alias{ps_match3}
\title{Propensity Score Matching}
\usage{
ps_match3(x)
}
\arguments{
\item{x}{is the input vector to the function}
}
\value{
A list with match_id and match_x
}
\description{
Propensity Score Matching
}
\examples{
set.seed(199)
x<-runif(10)
ps_match3(x)
}
