% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ps_matching}
\alias{ps_matching}
\title{Propensity Score}
\usage{
ps_matching(x)
}
\arguments{
\item{x}{\itemize{
\item A vector of numeric argument
}}
}
\value{
A list with two vectors (\code{match_id}, \code{match_x}).
}
\description{
Propensity Score
}
\examples{
set.seed(1231)
x <- runif(5)
ps_matching(x)
}
